refactor: extract shared CodeLLDB infrastructure into @debugmcp/codelldb-common (#324)
Move the CodeLLDB vendoring script, executable resolver, and binary-format
detector out of adapter-rust into a new workspace package so CodeLLDB-backed
adapters (rust today, cpp next: #325) share one vendored copy (~150MB/platform)
instead of each downloading their own.

- New packages/codelldb-common: vendor-codelldb.js (all CODELLDB_* env knobs
  unchanged), codelldb-resolver, binary-detector, and newly extracted spawn
  glue (codelldb-command.ts): --port/--liblldb argv, embedded-Python env
  scrub, Windows space-in-path mirror, and the console->terminal
  normalization (#223)
- adapter-rust keeps its import paths and public exports via re-export shims
  (package tests vi.mock those specifiers) and delegates buildAdapterCommand
  to the extracted glue; rust-specific env (DLLTOOL, RUST_BACKTRACE) stays
- Vendor path updated in all consumers: resolver monorepo fallbacks,
  adapter-policy-rust spawn fallback, check-adapters.js, bundle-cli.js
  CodeLLDB copy, .dockerignore, Dockerfile manifest COPYs, build-packages
  build order, prepare-pack workspace list, vitest alias

Zero behavior change: adapter-rust unit tests pass untouched (158),
full unit suite 3304 green, rust e2e smoke 4/4 against the relocated
vendor tree, pnpm vendor:status and test-bundle.cjs verified.
