+++ .dockerignore
@@ -73,13 +73,13 @@ packages/*/dist
 packages/*/tsconfig.tsbuildinfo
 packages/mcp-debugger/package
 packages/mcp-debugger/package/**
-packages/adapter-rust/vendor/codelldb/*
-!packages/adapter-rust/vendor
-!packages/adapter-rust/vendor/codelldb
-!packages/adapter-rust/vendor/codelldb/linux-x64
-!packages/adapter-rust/vendor/codelldb/linux-x64/**
-!packages/adapter-rust/vendor/codelldb/linux-arm64
-!packages/adapter-rust/vendor/codelldb/linux-arm64/**
+packages/codelldb-common/vendor/codelldb/*
+!packages/codelldb-common/vendor
+!packages/codelldb-common/vendor/codelldb
+!packages/codelldb-common/vendor/codelldb/linux-x64
+!packages/codelldb-common/vendor/codelldb/linux-x64/**
+!packages/codelldb-common/vendor/codelldb/linux-arm64
+!packages/codelldb-common/vendor/codelldb/linux-arm64/**
 build
 *.tsbuildinfo
 **/*.tsbuildinfo
