+++ vitest.config.ts
@@ -18,6 +18,7 @@ const sharedResolve = {
     // Keep project aliases pointing to TS sources
     { find: '@', replacement: path.resolve(__dirname, './src') },
     { find: '@debugmcp/shared', replacement: path.resolve(__dirname, './packages/shared/src/index.ts') },
+    { find: '@debugmcp/codelldb-common', replacement: path.resolve(__dirname, './packages/codelldb-common/src/index.ts') },
     { find: '@debugmcp/adapter-mock', replacement: path.resolve(__dirname, './packages/adapter-mock/src/index.ts') },
     { find: '@debugmcp/adapter-python', replacement: path.resolve(__dirname, './packages/adapter-python/src/index.ts') },
     { find: '@debugmcp/adapter-ruby', replacement: path.resolve(__dirname, './packages/adapter-ruby/src/index.ts') },
